Create README.md
This report has the analysis for the Uber dataset. The report is basically concentrated on where the Demand and Supply of Uber went wrong so that we can conclude and suggest solutions to the problems that were faced. Our main focus will be on the Demand (request for cab) and Supply ( how many requests were actually served), the location that has been facing the shortage of cabs and the timeslot for which maximum trips were cancelled or no cars were available. Dataset has following columns:-

1. Request id: The unique identification number given to the customer at the time he/she books a cab.

2. Pickup point: The location from where the person had been picked up for the ride (we have only 'Airport' and 'City' as locations over here.)

3. Driver id: Unique identification number for the driver.

4. Status: This column indicates the status of the trip i.e whether it was completed,cancelled etc.

5. Request timestamp: The date and time at which the cab was requested for.

6. Drop timestamp: The date and time at which the person was dropped to the destination.

This will be our Approach:-

1. Data Cleaning

2. Deriving new columns

3. Univariate Analysis, Bivariate Analysis and Multivariate Analysis

4. Final Conclusion